refactor: extract `Field`
